moved MockIAFWebViewDelegate into its own file
